spaces.Box expects a floating array, updated all cases where this was not true and modified float32 to float64 as float array default to float64. Otherwise space.Box raises warning that dtype precision (float32) is lower than array precision (float64).

+++ tests/spaces/test_utils.py
@@ -14,7 +14,7 @@ spaces = [
     Tuple(
         [
             Discrete(5),
-            Box(low=np.array([0, 0]), high=np.array([1, 5]), dtype=np.float32),
+            Box(low=np.array([0.0, 0.0]), high=np.array([1.0, 5.0]), dtype=np.float64),
         ]
     ),
     Tuple((Discrete(5), Discrete(2), Discrete(2))),
@@ -24,7 +24,7 @@ spaces = [
         {
             "position": Discrete(5),
             "velocity": Box(
-                low=np.array([0, 0]), high=np.array([1, 5]), dtype=np.float32
+                low=np.array([0.0, 0.0]), high=np.array([1.0, 5.0]), dtype=np.float64
             ),
         }
     ),
